Matrix, Boolean, Panel Dynamic Questions - The property 'commentText' is unavailable fixed #5562
