+++ src/question_baseselect.ts
@@ -1632,15 +1632,6 @@ Serializer.addClass(
   "selectbase",
   [
     { name: "showCommentArea:switch", layout: "row", visible: true, category: "general" },
-    {
-      name: "commentText",
-      dependsOn: "showCommentArea",
-      visibleIf: function (obj: any) {
-        return obj.hasComment;
-      },
-      serializationProperty: "locCommentText",
-      layout: "row",
-    },
     "choicesFromQuestion:question_selectbase",
     {
       name: "choices:itemvalue[]", uniqueProperty: "value",
