+++ src/question_expression.ts
@@ -432,8 +432,7 @@ Serializer.addClass(
     { name: "defaultValueExpression", visible: false },
     { name: "defaultValue", visible: false },
     { name: "correctAnswer", visible: false },
-    { name: "requiredIf", visible: false },
-    { name: "showCommentArea", visible: false }
+    { name: "requiredIf", visible: false }
   ],
   function() {
     return new QuestionExpressionModel("");
